Fix decompilation bugs: crashes, clickgui, fonts, AutoArmor/AutoCrystal, shader
Fonts: fix all-fonts-blank — integer division in glyph texcoords (FontHelper) → float.
UI sound: fix can't-switch — field shadowing (SettingFontsSearchHelper42.Inner.flag) → is92 hit-test.
clickgui enum dropdown: fix overlapping options — field shadowing (SearchHelper419.num) → index.
AutoArmor: fix crash-on-toggle — divide-by-zero in durability% (FireworksHelper/ArmorSearchHelper4) → float div.
AutoCrystal: align placement with original — is849 forceSuicide inversion; basePlace candidate-skip (break label99).

Crashes — cast-before-instanceof on render/tick threads (unprotected by try/catch):
- Tooltips (trapdoor→ShulkerBoxBlock), Highlight/Nuker/Feature_39 (EntityHitResult→BlockHitResult)
- AutoBowRelease (bow→CrossbowItem), AutoCrystal (crystal id reuse→SpawnTimeHelper_2), RocketExtender (firework id reuse→DuckFireworkEntity)
- Systematic sweep of ~120 cast-before-instanceof sites (same-type + negated-guard forms); packet-thread duck casts left as-is (protected by MixinClientConnection try/catch).

Shader ESP: clear shader framebuffers to transparent black (0,0,0,0) to prevent white leak under load.

+++ src/main/java/me/mioclient/ArmorSearchHelper4.java
@@ -176,6 +176,6 @@ public class ArmorSearchHelper4 implements SearchHelper_4 {
     }
 
     public static int get1905(ItemStack itemStack) {
-        return MathHelper.clamp(100 - ((int) Math.ceil((Float.intBitsToFloat(1065353216) - ((itemStack.getMaxDamage() - itemStack.getDamage()) / itemStack.getMaxDamage())) * Float.intBitsToFloat(1120403456))), 1, 100);
+        return MathHelper.clamp(100 - ((int) Math.ceil((Float.intBitsToFloat(1065353216) - ((itemStack.getMaxDamage() - itemStack.getDamage()) / (float) itemStack.getMaxDamage())) * Float.intBitsToFloat(1120403456))), 1, 100);
     }
 }
